% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C19Cases.R
\name{C19Cases}
\alias{C19Cases}
\title{COVID-19 Cases}
\usage{
C19Cases()
}
\value{
A data set consisting confirmed cases
}
\description{
Calling deaths and confirmed functions will return a data frame consisting of each countries total death and confirmed cases of COVID-19 by day
}
