% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C19Deaths.R
\name{C19Death}
\alias{C19Death}
\title{COVID-19 Deaths}
\usage{
C19Death()
}
\value{
A data set consisting death cases
}
\description{
Calling deaths and confirmed functions will return a data frame consisting of each countries total death and confirmed cases of COVID-19 by day
}
