% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C19WorldMap.R
\name{C19WorldMap}
\alias{C19WorldMap}
\title{COVID-19 cases and Deaths on World map}
\usage{
C19WorldMap(type, date)
}
\arguments{
\item{type}{1 is reserved for death cases and 2 is reserved for confirmed case}

\item{date}{A string consisting the specific date when the heatmep is desired to be drawn in m/d/yy format}
}
\value{
A heatmap of the world at the desired date
}
\description{
Calling DeathCasesMap will return a heatmap of the World, displaying each countries situation of death and confirmed cases
}
