% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19vsTime.R
\name{COVID19vsTime}
\alias{COVID19vsTime}
\title{COVID-19 cases and Death over time}
\usage{
COVID19vsTime(Country, Start, Stop)
}
\arguments{
\item{Country}{An string consisting the name a country}

\item{Start}{An string consisting the first date the plot needs to draw in yyyy-mm-dd format}

\item{Stop}{An string consisting the last date the plot needs to draw in yyyy-mm-dd format}
}
\value{
The evolution of deaths and confirmed cases over time
}
\description{
COVID-19 cases and Death over time
}
